Break down `githubApiService.ts` into several smaller files and move all GH related files to a dedicated folder
