+++ plugins/bulk-import-backend/src/github/GithubAppManager.test.ts
@@ -20,7 +20,7 @@ import { ScmIntegrations } from '@backstage/integration';
 import type { RestEndpointMethodTypes } from '@octokit/rest';
 import { DateTime } from 'luxon';
 
-import type { ExtendedGithubCredentialsProvider } from '../types';
+import type { ExtendedGithubCredentialsProvider } from './types';
 import {
   CustomGithubCredentialsProvider,
   CustomSingleInstanceGithubCredentialsProvider,
