+++ plugins/bulk-import-backend/src/github/index.ts
@@ -1,3 +1,4 @@
+
 /*
  * Copyright 2024 The Janus IDP Authors
  *
@@ -14,6 +15,6 @@
  * limitations under the License.
  */
 
-export * from './auth';
-export * from './loggingUtils';
-export * from './pagination';
+export * from './types';
+export * from './GithubAppManager';
+export * from './githubApiService';
